+++ CHANGELOG.md
@@ -7,11 +7,15 @@ Add changes to the "Unreleased Changes" section. Once you create a version (and 
 to a section with the version name.
 
 ## Unreleased Changes
+
+## 0.48.0
+
 * Prevent manual submission of `Sales Invoice Additional Fields` where user can submit with the intent of resending it ([Issue](https://github.com/lavaloon-eg/ksa_compliance/issues/233))
   * It should only be submitted by the hourly job or the sync page.
 * Display `ZATCA Integration Status` as Intro in the `Sales Invoice`
 
 * Add revocation option for zatca business settings and CSID.
+
 ## 0.47.0
 
 * Support displaying `Return Against Additional References` in `ZATCA Phase 2 print format`.